requirements.txt
python-telegram-bot==20.3
python-dotenv
openai
pytz
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2